** Routing 별로 Workstation Info Definition Save / Change

+++ APS.Database/dbo/Tables/DashBoard/WorkStation.sql
@@ -4,8 +4,8 @@
 	[Title] NVarChar(20) Not Null,
 	[Image] NVarChar(55) Null Default('DefaultImage'),
 	[Description] NVarChar(255) Not Null,
-	[SetupTime] INT NOT NULL,
-	[ProcessingTime] INT NOT NULL,
+	[SetupTime] float NOT NULL,
+	[ProcessingTime] float NOT NULL,
 
 	[ViewOrder] Int	Default(1),
 	[GroupUID] Int Not Null Default(1) -- 작성 그룹 코드번호
